backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
python-multipart>=0.0.9,<1.0
Pillow>=10.4,<11.0
pdf2image>=1.17,<2.0
easyocr>=1.7,<2.0
pythainlp>=5.0,<6.0
pypdfium2>=5.6,<6.0
nltk>=3.9,<4.0
